requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APIResponse==0.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
authclient==1.0
black==23.7.0
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
crashtest==0.4.1
credentials==1.1
decorator==5.1.1
distlib==0.3.7
Django==4.2.4
dulwich==0.21.5
exceptiongroup==1.1.3
fastapi==0.101.0
filelock==3.12.2
flake8==5.0.4
Flask==2.3.2
frozenlist==1.4.0
h11==0.14.0
html5lib==1.1
idna==2.10
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
openapi-client==1.1.7
packaging==23.1
pathspec==0.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pyflakes==2.5.0
PyJWT==2.7.0
pyproject_hooks==1.0.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rapidfuzz==2.15.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
self==2020.12.3
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
tox==3.28.0
trove-classifiers==2023.8.7
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
urllib3>=1.26.0,<2.0.0
uvicorn==0.23.2
virtualenv>=20.22.0,<21.0.0
webencodings==0.5.1
Werkzeug==2.3.6
xattr==0.10.1
yarl==1.9.2
zipp==3.16.2
